+++ .gitignore
@@ -1,3 +1,4 @@
 /venv
 /.idea
 */__pycache__/*
+/data
